backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
pydantic
python-multipart
pdfplumber
python-docx
